[AN-354] Windows adaptation (!50)
Reviewed-on: https://gitea.devos.club/antiq/typescript-llvm/pulls/50
Co-Committed-By: eugene.makarov <eugene.makarov@myoffice.team>

+++ test/unit/cpp_integration/declarations/printable.d.ts
@@ -2,12 +2,12 @@
 
 declare module "cpp" {
     export class Printable {
-        constructor(point: Point, s: string, i: int64_t);
+        constructor(point: Point, s: string, i: int32_t);
 
         @ValueType
         private point_: Point;
         @ValueType
         private string_: string;
-        private i_: int64_t;
+        private i_: int32_t;
     }
 }
